Refactor AuthPageLayout
